feat:products&contacts refactor & add fieldByCode
