+++ packages/plugin-contacts-api/src/coc/utils.ts
@@ -557,40 +557,3 @@ export class CommonBuilder<IListArgs extends ICommonListArgs> {
     };
   }
 }
-
-export const customFieldsDataByFieldCode = async (object, subdomain) => {
-  const customFieldsData =
-    object.customFieldsData && object.customFieldsData.toObject
-      ? object.customFieldsData.toObject()
-      : object.customFieldsData || [];
-
-  const fieldIds = customFieldsData.map(data => data.field);
-
-  const fields = await sendFormsMessage({
-    subdomain,
-    action: 'fields.find',
-    data: {
-      query: {
-        _id: { $in: fieldIds }
-      }
-    },
-    isRPC: true,
-    defaultValue: []
-  });
-
-  const fieldCodesById = {};
-
-  for (const field of fields) {
-    fieldCodesById[field._id] = field.code;
-  }
-
-  const results: any = {};
-
-  for (const data of customFieldsData) {
-    results[fieldCodesById[data.field]] = {
-      ...data
-    };
-  }
-
-  return results;
-};
