+++ packages/plugin-contacts-api/src/utils.ts
@@ -29,7 +29,8 @@ import {
 import { companySchema } from './models/definitions/companies';
 import { ICustomField, ILink } from '@erxes/api-utils/src/types';
 import { fetchEs } from '@erxes/api-utils/src/elasticsearch';
-import { customFieldsDataByFieldCode } from './coc/utils';
+import { customFieldsDataByFieldCode } from '@erxes/api-utils/src/fieldUtils';
+import { sendCommonMessage } from './messageBroker';
 
 const EXTEND_FIELDS = {
   CUSTOMER: [
@@ -92,7 +93,8 @@ export const findCustomer = async (
   if (customer) {
     customer.customFieldsDataByFieldCode = await customFieldsDataByFieldCode(
       customer,
-      subdomain
+      subdomain,
+      sendCommonMessage
     );
   }
 
@@ -180,7 +182,8 @@ export const findCompany = async (
   if (company) {
     company.customFieldsDataByFieldCode = await customFieldsDataByFieldCode(
       company,
-      subdomain
+      subdomain,
+      sendCommonMessage
     );
   }
