+++ packages/plugin-products-api/src/graphql/schema/product.ts
@@ -52,6 +52,7 @@ export const types = (tagsAvailable, contactsAvailable) => `
     unitPrice: Float
     categoryId: String
     customFieldsData: JSON
+    customFieldsDataByFieldCode: JSON
     createdAt: Date
     ${tagsAvailable ? `getTags: [Tag]` : ''}
     tagIds: [String]
